+++ apps/server/utils/indexes.dart
@@ -23,6 +23,20 @@ db.restaurants.createIndex(
 1 == ascending
 -1 == descending
 
+
+TODO: also include search index
+[
+  {
+    $search: {
+      index: "title",
+      text: {
+        query: "string",
+        path: "string"
+      }
+    }
+  }
+]
+
  */
 main() async {
   final db = await Db.create(TestHelper.local);
@@ -131,12 +145,13 @@ main() async {
 
   /// basic
   await db.ensureIndex(TicketService.collectionName,
-      name: 'firmId_chainId_boutiqueId_isDeleted_nonUniqueId_creationDate',
+      name: 'firmId_chainId_boutiqueId_isDeleted_nonUniqueId_contactId_isDeleted_creationDate',
       keys: {
         'firmId': 1,
         'chainId': 1,
         'boutiqueId': 1,
         'nonUniqueId': 1,
+        'contactId': 1,
         'isDeleted': 1,
         'creationDate': -1
       });