+++ packages/article_service/lib/src/article_service_base.dart
@@ -16,7 +16,7 @@ abstract class _Helpers {
       where
           .eq('firmId', firmId)
           .eq('chainId', request.chainId)
-          .eq('creationDate', request.category.creationDate);
+          .eq('category.creationDate', request.category.creationDate);
 
   static SelectorBuilder selectPhoto(String firmId, PhotoRequest request) =>
       where
@@ -205,7 +205,7 @@ class ArticleService extends ArticleServiceBase {
           .eq('firmId', userPermission.firmId)
           .eq('chainId', request.chainId);
 
-      final bool isDeviceResync = request.lastFetchTimestampUTC.hasSeconds();
+      final bool isDeviceResync = request.lastFetchTimestampUTC.isNotEmpty;
       final idsSet = <int>{};
       if (isDeviceResync) {
         final documents = await collectionArticle.find(selector).toList();
@@ -218,7 +218,7 @@ class ArticleService extends ArticleServiceBase {
       }
       final list = await collectionArticle.find(selector).toList();
       if (list.isEmpty) {
-        return CalibresResponse.create();
+        return CalibresResponse(ids:idsSet);
       }
       final calibres = <CalibrePb>[];
       for (final e in list) {
@@ -254,7 +254,6 @@ class ArticleService extends ArticleServiceBase {
           'user cannot access data from chain ${request.chainId}');
     }
 
-    // TODO: search by calibreId
     try {
       final selector = where
           .eq('firmId', userPermission.firmId)
@@ -364,6 +363,8 @@ class ArticleService extends ArticleServiceBase {
         ..userId = userPermission.userId
         ..lastTouchTimestampUTC = DateTime.now().toUtc().timestampProto;
 
+//      final resultFindOne = await collectionCategory.findOne(_Helpers.selectCategory(userPermission.firmId, request));
+
       final result = await collectionCategory.replaceOne(
           _Helpers.selectCategory(userPermission.firmId, request),
           categoryMongo.toProto3Json() as Map<String, dynamic>,
@@ -437,9 +438,11 @@ class ArticleService extends ArticleServiceBase {
       final selector = SelectorBuilder()
           .eq('firmId', userPermission.firmId)
           .eq('chainId', request.chainId);
+
       final list = await collectionCategory.find(selector).toList();
       final categories = <CategoryPb>[];
       for (final e in list) {
+      
         final categoryMongo = CategoryMongo.create()
           ..mergeFromProto3Json(e, ignoreUnknownFields: true);
         categories.add(categoryMongo.category);
@@ -458,7 +461,7 @@ class ArticleService extends ArticleServiceBase {
 
   @override
   Future<CategoryPb> readOneCategory(
-      ServiceCall call, FindCategoryRequest request) async {
+      ServiceCall? call, FindCategoryRequest request) async {
     _db.isConnected ? null : await _db.open();
     final userPermission = isTest
         ? userPermissionIfTest ?? UserPermissions()
@@ -475,8 +478,14 @@ class ArticleService extends ArticleServiceBase {
     try {
       final selector = SelectorBuilder()
           .eq('firmId', userPermission.firmId)
-          .eq('chainId', request.chainId)
-          .eq('title', request.title);
+          .eq('chainId', request.chainId);
+
+      if (request.title.isNotEmpty) {
+        selector.match('title', '^${RegExp.escape(request.title)}*');
+      }
+
+      // pipeline is a work in progress
+
       final category = await collectionCategory.findOne(selector);
       if (category != null) {
         final categoryMongo = CategoryMongo.create()
@@ -751,7 +760,7 @@ class ArticleService extends ArticleServiceBase {
       final selector = SelectorBuilder()
           .eq('firmId', userPermission.firmId)
           .eq('chainId', request.chainId);
-      if (request.lastFetchTimestampUTC.hasSeconds()) {
+      if (request.lastFetchTimestampUTC.isNotEmpty) {
         selector.and(where.gte('lastTouchTimestampUTC',
             request.lastFetchTimestampUTC.toDateTime()));
       }