+++ packages/ticket_service/lib/src/ticket_service_base.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:fence_service/mongo_dart.dart' hide Timestamp;
 import 'package:fence_service/fence_service.dart';
 import 'package:fence_service/grpc.dart';
@@ -63,6 +61,7 @@ class TicketService extends TicketServiceBase {
         ..chainId = request.ticket.counterfoil.chainId
         ..firmId = userPermission.firmId
         ..userId = userPermission.userId
+        ..contactId = request.ticket.contactId
         ..lastTouchTimestampUTC = DateTime.now().toUtc().timestampProto;
 
       final result = await collection
@@ -131,7 +130,7 @@ class TicketService extends TicketServiceBase {
       selector.and(where.eq('boutiqueId', request.boutiqueId));
     }
 
-    if (request.lastFetchTimestampUTC.hasSeconds()) {
+    if (request.lastFetchTimestampUTC.isNotEmpty) {
       selector.and(where.gte(
           'lastTouchTimestampUTC', request.lastFetchTimestampUTC.toDateTime()));
     }
@@ -378,6 +377,7 @@ class TicketService extends TicketServiceBase {
         ..chainId = ticketPb.counterfoil.chainId
         ..firmId = userPermission.firmId
         ..userId = userPermission.userId
+        ..contactId = ticketPb.contactId
         ..lastTouchTimestampUTC = nowTimestampUtc;
       ticketsMap.add(ticketMongo.toProto3Json() as Map<String, dynamic>);
     }